Yet another patch by Sven Eckelmann:
Don't truncate file descriptor in use_kernel_module

We should not truncate the file descriptor of /dev/batgat from int32_t
to int8_t (return type of use_kernel_module) when we save it as int32_t
in batman_if::udp_tunnel_sock.




git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1171 45894c77-fb22-0410-b583-ff6e7d5dbf6c

+++ bsd/kernel.c
@@ -143,7 +143,7 @@ int8_t use_kernel_module( char *dev )
 	return -1;
 }
 
-int8_t use_gateway_module(void)
+int32_t use_gateway_module(void)
 {
 	return -1;
 }
