+++ linux/kernel.c
@@ -193,7 +193,7 @@ int8_t bind_to_iface( int32_t sock, char *dev ) {
 
 
 
-int8_t use_gateway_module(void) {
+int32_t use_gateway_module(void) {
 
 	int32_t fd;
