+++ os.h
@@ -70,7 +70,7 @@ int32_t get_send_redirects( char *dev );
 void set_forwarding( int32_t state );
 int32_t get_forwarding( void );
 int8_t bind_to_iface( int32_t sock, char *dev );
-int8_t use_gateway_module(void);
+int32_t use_gateway_module(void);
 
 /* posix.c */
 void print_animation( void );
